requirements.txt
pillow==7.2.0
pywin32==228